added oracle dep in dependencies section and removed the jar from the lib dir
- updated the version to 1.2
- also added a mavenRepoUrl to the gradle.properties_template

So the thought here is that the oracle jar will be listed as a top level dependency and picked up as a transitive dependency of this artifact.  This requires a private mavenRepo with an oracle jdbc jar that conforms to the naming convention.  The mavenRepoUrl should be set to this private mavenRepo.